+++ src/Core/BoundedContexts/CAPI/WB.Core.BoundedContexts.CAPI/Views/InterviewDetails/LinkedAnswerViewModel.cs
@@ -4,13 +4,13 @@ namespace WB.Core.BoundedContexts.Capi.Views.InterviewDetails
 {
     public class LinkedAnswerViewModel : Cirrious.MvvmCross.ViewModels.MvxViewModel, ICloneable
     {
-        public LinkedAnswerViewModel(int[] propagationVector, string title)
+        public LinkedAnswerViewModel(decimal[] propagationVector, string title)
         {
             this.PropagationVector = propagationVector;
             this.Title = title;
         }
 
-        public int[] PropagationVector { get; private set; }
+        public decimal[] PropagationVector { get; private set; }
         public string Title { get; private set; }
 
         public object Clone()
