+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Commands/Interview/AnswerGeoLocationQuestionCommand.cs
@@ -13,7 +13,7 @@ namespace WB.Core.SharedKernels.DataCollection.Commands.Interview
         public double Longitude { get; private set; }
         public double Accuracy { get; private set; }
 
-        public AnswerGeoLocationQuestionCommand(Guid interviewId, Guid userId, Guid questionId, int[] rosterVector,
+        public AnswerGeoLocationQuestionCommand(Guid interviewId, Guid userId, Guid questionId, decimal[] rosterVector,
             DateTime answerTime, double latitude, double longitude, double accuracy, DateTimeOffset timestamp)
             : base(interviewId, userId, questionId, rosterVector, answerTime)
         {
