+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/FlagRemovedFromAnswer.cs
@@ -5,7 +5,7 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview
 {
     public class FlagRemovedFromAnswer : QuestionActiveEvent
     {
-        public FlagRemovedFromAnswer(Guid userId, Guid questionId, int[] propagationVector)
+        public FlagRemovedFromAnswer(Guid userId, Guid questionId, decimal[] propagationVector)
             : base(userId, questionId, propagationVector) {}
     }
 }
