+++ src/Tests/WB.Core.SharedKernels.DataCollection.Tests/InterviewTests/when_answer_on_question_triggers_validation_evaluation_of_unanswered_question.cs
@@ -52,7 +52,7 @@ namespace WB.Core.SharedKernels.DataCollection.Tests.InterviewTests
         };
 
         Because of = () =>
-              interview.AnswerNumericIntegerQuestion(userId, answeredQuestionId, new int[] { }, DateTime.Now, 1);
+              interview.AnswerNumericIntegerQuestion(userId, answeredQuestionId, new decimal[] { }, DateTime.Now, 1);
 
         It should_raise_AnswerDeclaredValid_event = () =>
              eventContext.ShouldContainEvent<AnswerDeclaredValid>(@event
