+++ src/Tests/WB.Core.SharedKernels.DataCollection.Tests/InterviewTests/when_validation_expression_execution_throw_exception_for_question_in_invalid_state.cs
@@ -62,7 +62,7 @@ namespace WB.Core.SharedKernels.DataCollection.Tests.InterviewTests
         };
 
         Because of = () =>
-            interview.AnswerNumericRealQuestion(userId, validatingQuestionId, new int[] { }, DateTime.Now, 0);
+            interview.AnswerNumericRealQuestion(userId, validatingQuestionId, new decimal[] { }, DateTime.Now, 0);
 
         It should_not_raise_AnswerDeclaredValid_event_with_QuestionId_equal_to_validatingQuestionId = () =>
             eventContext.ShouldNotContainEvent<AnswerDeclaredValid>(@event
